Add script to handle likes
