+++ app/javascript/packs/posts.js
@@ -1,0 +1,11 @@
+$(function() {
+    $(".post-like").on("click", function() {
+        var post_id = $(this).data("id");
+        $.ajax({
+            url: "/post/like/"+post_id,
+            method: 'GET'
+            }).done(function(response) {
+                // console.log(response);
+            })
+    });
+});
