--- app/javascript/test.js
@@ -1,22 +1,0 @@
- 
-
-    // console.log("file changed");
-    // function readURL(input) {
-
-    //     if (input.files && input.files[0]) {
-    //         var reader = new FileReader();
-    
-    //         reader.onload = function (e) {
-    //             $('#blah').attr('src', e.target.result);
-    //         }
-    
-    //         reader.readAsDataURL(input.files[0]);
-    //     }
-    // }
-
-    // $("body").on('change',"#imgUp", function() {
-    //     console.log('clicked');
-    // })
-//  function test() {
-//         console.log('clicked');
-// }
